requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
SQLAlchemy==2.0.34
alembic==1.13.2
pydantic==2.9.1
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
aiosmtplib==3.0.2
email-validator==2.2.0
psycopg2-binary==2.9.9
psycopg2
